remove all from App.js
